create plane using dip and dip direction angles
plane can be created using point, dip and dip direction angles. this is going to be very useful for Geoscientists and structural geologists. dip angle is the maximum inclination to the horizontal and dip direction angle is the direction direction of the horizontal trace of the line of dip, measured clockwise from north. more information about dip and dip direction angles can be found in here: http://www.structuralgeology.org/2009/11/strike-dip-right-hand-rule.html