Change expandPolygon call to include 'cleanuploops', true
The picture of the polygon demoExpandPolygon_03.png shows the interior
polygons as two separate loops.  Passing the 'cleanuploops', true to
expandPolygon draws the ploygons this way.  Omitting the 'cleanuploops',
true makes expandPolygon create a third polygon triangle with it's
hypotenuse an extension of the hypotenuse of the top triangle and it's
sides extensions of the top of the bottom polygon and the vertical side
of the top triangle, with both sides meeting at a right angle.  This
would be ok too, but it is not what is drawn in demoExpandPolygon_03.png.