use current axes handle if no axes is provided
This solved the issue when drawPolyline3d is used with the app designer. The function will create a new empty figure even if the axes handle is provided in the first input of the function. To solve this problem, check if the first input is the axes handle, if not create a new figure. This also seems to be true for other functions in matGeom, however, this one is tested.